Made a few changes to make building jars possible.
