+++ Utilities/ErrorHandler.java
@@ -33,20 +33,13 @@ import java.io.StringWriter;
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-public class ErrorHandler {
+class ErrorHandler {
 
     static void fileNotFound(String filename) {
         Dialogs.errorDialog(
                 "File not found",
                 "File not found",
                 "The specified file '" + filename + "' was not found.\nPlease include full path to the file.");
-    }
-
-    public static void invalidDateTimeFormat() {
-        Dialogs.errorDialog(
-                "Invalid date/time format",
-                "Invalid date/time format",
-                "Please enter a valid date/time format. \nExample: yyyy/MM/dd HH:mm:ss");
     }
 
     /**
